Much UNIX. Very pipe. So stream. Wow.
The race runner now logs races to the ./log directory by just appending
all communications from the server into a timestamped file. The
visualiser is now a stand-alone program that starts a browser and pipes
its newline-split stdin into a websocket.

To vis a race that has already happened, `./visualiser.ls <
logs/whatever`.

To real-time vis a race that's happening, first start the race (which
will start writing a file), then

    tail -n +0 -F logs/whatever | ./visualiser

That `-F` to `tail` is for *follow*, which uses inotify to print more
lines to `stdout` as they're written to the file.  `-n +0` just ensures
it `tail`s not only the last 10 lines but everything from the start of
the file.

`vis-latest` is a handy script that just picks the latest log and runs
`visualiser.ls` on it.

The logs aren't checked in.