Simplify send code
Some of that logic for figuring out what data exactly to pass back from
the AI seemed flaky and was certainly awkward to read. Similarly when
calling `send.apply`.

Some other minor refactorings.

Also increased the (nonsensically) constant throttle to `0.6`, so we're
now like a second faster around the track. :)